Another warning fix
git-svn-id: https://svn.osgeo.org/gdal/trunk@25254 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/openjpeg/openjpegdataset.cpp
@@ -1759,7 +1759,7 @@ GDALDataset * JP2OpenJPEGDataset::CreateCopy( const char * pszFilename,
 
             /* Now let's extend the jp2c box header so that the */
             /* res box becomes a sub-box of it */
-            nLBoxJP2H += poRes->GetDataLength() + 8;
+            nLBoxJP2H += (int)poRes->GetDataLength() + 8;
             nLBoxJP2H = CPL_MSBWORD32( nLBoxJP2H );
             VSIFSeekL(fp, 32, SEEK_SET);
             VSIFWriteL(&nLBoxJP2H, 1, 4, fp);
